+++ frontend/features/artist/hooks/index.ts
@@ -2,3 +2,4 @@ export { useArtistData } from "./useArtistData";
 export { useArtistActions } from "./useArtistActions";
 export { useDownloadActions } from "./useDownloadActions";
 export { useArtistTracks } from "./useArtistTracks";
+export { useProviderArtistFallback } from "./useProviderArtistFallback";